# api/openapi.yaml
openapi: 3.0.3
info:
  title: FitCore API
  version: 0.1.0
  description: >-
    Gym management platform covering branches, members, memberships, group
    classes, bookings, attendance check-in/out, invoicing, staff and trainers.
    Contract-first: routes are documented ahead of their HTTP adapters, which
    are implemented incrementally. Every operation except health, metrics and
    token issuance requires a short-lived Bearer access token (JWT) carrying
    the caller's permission set; refresh tokens rotate via /auth/refresh.
security:
  - BearerAuth: []
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /members/{memberId}/attendance:
    parameters:
      - $ref: "#/components/parameters/MemberID"
    get:
      tags: [attendance]
      summary: List a member's attendance records
      operationId: listMemberAttendance
      x-permission: attendance:read
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
        "500":
          $ref: "#/components/responses/InternalError"
  /members/{memberId}/invoices:
    parameters:
      - $ref: "#/components/parameters/MemberID"
    get:
      tags: [billing]
      summary: List a member's invoices
      operationId: listMemberInvoices
      x-permission: billing:read
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /classes/{classId}/bookings:
    parameters:
      - $ref: "#/components/parameters/ClassID"
    get:
      tags: [bookings]
      summary: List bookings for a class
      operationId: listClassBookings
      x-permission: bookings:read
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "500":
          $ref: "#/components/responses/InternalError"
  /bookings:
    post:
      tags: [bookings]
      summary: Book a seat in a class
      description: Fails if the class is at capacity or already fully booked.
      operationId: createBooking
      x-permission: bookings:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCreateRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /bookings/{bookingId}:
    parameters:
      - $ref: "#/components/parameters/BookingID"
    get:
      tags: [bookings]
      summary: Get a booking by ID
      operationId: getBooking
      x-permission: bookings:read
      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"
  /bookings/{bookingId}/cancel:
    parameters:
      - $ref: "#/components/parameters/BookingID"
    post:
      tags: [bookings]
      summary: Cancel a booking
      description: Frees the class seat; freezes the member's quota.
      operationId: cancelBooking
      x-permission: bookings:update
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /attendance/check-in:
    post:
      tags: [attendance]
      summary: Check a member in
      description: Requires an active membership for the member at the branch.
      operationId: checkIn
      x-permission: attendance:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceCheckInRequest"
      responses:
        "201":
          description: Open attendance record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /attendance/check-out:
    post:
      tags: [attendance]
      summary: Check a member out
      description: Closes the member's open attendance record.
      operationId: checkOut
      x-permission: attendance:update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceCheckOutRequest"
      responses:
        "200":
          description: Closed attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /attendance/{attendanceId}:
    parameters:
      - $ref: "#/components/parameters/AttendanceID"
    get:
      tags: [attendance]
      summary: Get an attendance record by ID
      operationId: getAttendance
      x-permission: attendance:read
      responses:
        "200":
          description: Attendance record found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "404":
          $ref: "#/components/responses/NotFound"
  /invoices:
    post:
      tags: [billing]
      summary: Create an invoice
      operationId: createInvoice
      x-permission: billing:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceCreateRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /invoices/{invoiceId}:
    parameters:
      - $ref: "#/components/parameters/InvoiceID"
    get:
      tags: [billing]
      summary: Get an invoice by ID
      operationId: getInvoice
      x-permission: billing:read
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [billing]
      summary: Update an invoice
      description: Marks an invoice paid, failed or void.
      operationId: updateInvoice
      x-permission: billing:update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdateRequest"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Short-lived access token issued by POST /auth/login, signed with
        HMAC-SHA256. Claims include sub (staff id), jti, perms (permission
        set) and exp.
  parameters:
...
    MemberID:
      name: memberId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
    ClassID:
      name: classId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    BookingID:
      name: bookingId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    AttendanceID:
      name: attendanceId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    InvoiceID:
      name: invoiceId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
    BadRequest:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with the current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    UUID:
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    Timestamp:
      type: string
      format: date-time
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message, safe for clients.
        code:
          type: string
          nullable: true
          description: Stable machine-readable error code.
      example:
        error: "member not found"
        code: "member_not_found"
...
    BookingStatus:
      type: string
      enum: [booked, cancelled]
    Booking:
      type: object
      required: [id, classId, memberId, status, bookedAt, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        classId:
          $ref: "#/components/schemas/UUID"
        memberId:
          $ref: "#/components/schemas/UUID"
        status:
          $ref: "#/components/schemas/BookingStatus"
        bookedAt:
          $ref: "#/components/schemas/Timestamp"
        cancelledAt:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
          nullable: true
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    BookingCreateRequest:
      type: object
      required: [classId, memberId]
      properties:
        classId:
          $ref: "#/components/schemas/UUID"
        memberId:
          $ref: "#/components/schemas/UUID"
    Attendance:
      type: object
      required: [id, memberId, branchId, checkedInAt, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        memberId:
          $ref: "#/components/schemas/UUID"
        branchId:
          $ref: "#/components/schemas/UUID"
        membershipId:
          allOf:
            - $ref: "#/components/schemas/UUID"
          nullable: true
        checkedInAt:
          $ref: "#/components/schemas/Timestamp"
        checkedOutAt:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
          nullable: true
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    AttendanceCheckInRequest:
      type: object
      required: [memberId, branchId]
      properties:
        memberId:
          $ref: "#/components/schemas/UUID"
        branchId:
          $ref: "#/components/schemas/UUID"
    AttendanceCheckOutRequest:
      type: object
      required: [memberId]
      properties:
        memberId:
          $ref: "#/components/schemas/UUID"
    InvoiceStatus:
      type: string
      enum: [pending, paid, failed, void]
    Invoice:
      type: object
      required:
        - id
        - memberId
        - membershipId
        - amountCents
        - currency
        - status
        - dueAt
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        memberId:
          $ref: "#/components/schemas/UUID"
        membershipId:
          $ref: "#/components/schemas/UUID"
        amountCents:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        dueAt:
          $ref: "#/components/schemas/Timestamp"
        paidAt:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
          nullable: true
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    InvoiceCreateRequest:
      type: object
      required: [memberId, membershipId, amountCents, currency, dueAt]
      properties:
        memberId:
          $ref: "#/components/schemas/UUID"
        membershipId:
          $ref: "#/components/schemas/UUID"
        amountCents:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        dueAt:
          $ref: "#/components/schemas/Timestamp"
    InvoiceUpdateRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        dueAt:
          $ref: "#/components/schemas/Timestamp"
...